network: separate the parser part of v1_directives
In the v2 API of the device gateway server, directives must be received
from /v2/events as well as /v1/directives.

Therefore, the parsing part in v1_directives.cc is separated into a
separate module so that the module can be reused in later v2_events
implementations.
